+++ src/components/scorecard-header/scorecard-header.component.js
@@ -2,9 +2,9 @@ const skill = require('kcapp-bot/bot-skill');
 
 module.exports = {
     onCreate(input) {
-        var player = input.player;
+        const player = input.player;
 
-        var name = input.players[player.player_id].name;
+        let name = input.players[player.player_id].name;
         if (player.player.is_bot) {
             if (player.bot_config.player_id) {
                 name = `${name} as ${input.players[player.bot_config.player_id].name}`;
@@ -25,7 +25,7 @@ module.exports = {
             $("iframe").ready(function() {
                 setTimeout(() => {
                     // Give the video a few seconds to load
-                    var iframe = document.getElementById(`iframe-player-cam-${this.state.player.player_id}`);
+                    const iframe = document.getElementById(`iframe-player-cam-${this.state.player.player_id}`);
                     iframe.contentWindow.postMessage( { "style" : this.state.player.player.board_stream_css }, '*');
                 }, 2000);
             }.bind(this));