+++ src/components/scorecard/scorecard.component.js
@@ -13,6 +13,7 @@ const fourTwenty = require('./components/four_twenty');
 const killBull = require('./components/kill_bull');
 const gotcha = require('./components/gotcha');
 const jdcPractice = require('./components/jdc_practice');
+const knockout = require("./components/knockout");
 
 const types = require("./components/match_types");
 
@@ -20,7 +21,7 @@ const DART_CONTAINER_MAP = { 1: 'first', 2: 'second', 3: 'third' };
 
 module.exports = {
     onCreate(input) {
-        var player = input.player;
+        const player = input.player;
         this.state = {
             uuid: input.uuid,
             players: input.players,
@@ -143,6 +144,9 @@ module.exports = {
                 case types.JDC_PRACTICE:
                     jdcPractice.removeLast.bind(this)(dart, external);
                     break;
+                case types.KNOCKOUT:
+                    knockout.removeLast.bind(this)(dart, external);
+                    break;
             }
             dart.reset();
         } else {
@@ -196,6 +200,9 @@ module.exports = {
                 case types.JDC_PRACTICE:
                     submitting = jdcPractice.confirmThrow.bind(this)(external);
                     break;
+                case types.KNOCKOUT:
+                    submitting = knockout.confirmThrow.bind(this)(external);
+                    break;
             }
         }
         return submitting;
@@ -218,9 +225,9 @@ module.exports = {
     },
 
     getPayload() {
-        var first = this.getComponent(DART_CONTAINER_MAP[1]);
-        var second = this.getComponent(DART_CONTAINER_MAP[2]);
-        var third = this.getComponent(DART_CONTAINER_MAP[3]);
+        const first = this.getComponent(DART_CONTAINER_MAP[1]);
+        const second = this.getComponent(DART_CONTAINER_MAP[2]);
+        const third = this.getComponent(DART_CONTAINER_MAP[3]);
         return {
             player_id: this.state.playerId,
             leg_id: this.state.leg.id,