+++ src/pages/leg/components/leg/leg.component.js
@@ -405,4 +405,3 @@ module.exports = {
         this.state.isPlayerBoardCam = enabled;
     }
 };
-